TASK-2: Adding flask-script manager to initdb and db command to MigrateCommand
